Some touches
- Wykonano ostylowanie dla aktywnych przycisków w headerze strony My library

-  Odrobinę zrefaktoryzowano komponent paginacji

- Poprawnie rozwiązano konflikty poprzednich zmergowanych gałęzi
