+++ src/components/headerLibrary/headerLibrary.html
@@ -23,7 +23,7 @@
       <button class="btn btn__watched" data-display-selector="watched">
         Watched
       </button>
-      <button class="btn btn__queue" data-display-selector="queue">
+      <button class="btn btn--active btn__queue" data-display-selector="queue">
         Queue
       </button>
     </div>
