+++ src/components/headerLibrary/headerLibrary.scss
@@ -28,7 +28,7 @@
   width: 130px;
   height: 44px;
   cursor: pointer;
-
+  transition: ease-out 0.3s;
   &:last-child {
     margin-left: 20px;
   }
@@ -43,7 +43,7 @@
     border: none;
     box-shadow: 0px 8px 43px rgba(255, 107, 1, 0.6);
   }
-  &__active {
+  &--active {
     background: #ff6b08;
     border: none;
     box-shadow: 0px 8px 43px rgba(255, 107, 1, 0.6);
